FocusGuard/requirements.txt
opencv-python==4.8.1.78
mediapipe==0.10.8
python-vlc==3.0.20123